IRInterpreter compiler: Reject all vec2ops where the prefix is unknown while compiling

+++ Core/HLE/sceKernelInterrupt.cpp
@@ -884,7 +884,7 @@ static int sysclib_sprintf(u32 dst, u32 fmt) {
 
 	DEBUG_LOG(SCEKERNEL, "sysclib_sprintf result string has length %d, content:", (int)result.length());
 	DEBUG_LOG(SCEKERNEL, "%s", result.c_str());
-	if (!Memory::IsValidRange(dst, result.length() + 1)) {
+	if (!Memory::IsValidRange(dst, (u32)result.length() + 1)) {
 		ERROR_LOG(SCEKERNEL, "sysclib_sprintf result string is too long or dst is invalid");
 		return 0;
 	}
